ENH: Serialize device access.
The device lock (mutex) is grabbed before every access to the device.
For python threading this is normally not neccessary, as C functions are
called with the GIL held, but for grabFrame the GIL is released when
each row is read in so as to increase interactivity when used in a
threaded environment. Thus the lock prevents any other function from
accessing the device until the entire frame has been read.